pyproject.toml
[tool.poetry]
name = "ollama_pr_assistant"
version = "0.1.0"
description = "Offline AI-powered PR assistant using Ollama for code review, commit messages, test generation, and documentation."
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.11"

# Add your core dependencies
click = "^8.1.3"           # Optional CLI enhancement
python-dotenv = "^1.0.0"   # For reading .env
rich = "^13.4.2"            # Nice CLI formatting

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
black = "^24.3.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
